moved dagger module for being devfest independent
