+++ app/src/hh2016/java/de/devfest/ApplicationModule.java
@@ -1,4 +1,4 @@
-package de.devfest.injection;
+package de.devfest;
 
 import android.app.Application;
